requirements.txt
sounddevice
numpy
torch
torchaudio
scipy
TTS
ffmpeg
google-generativeai